Improved data extraction strategy

+++ TableTennis.DataAccess/Models/EndedEventsPage.cs
@@ -39,7 +39,7 @@ namespace TableTennis.DataAccess.Models
             {
                 public int Id { get; set; }
                 public string Name { get; set; }
-                public long? ImageId { get; set; }
+                public int? ImageId { get; set; }
                 public string Cc { get; set; }
             }
         }
